requirements.txt
langchain==0.0.225
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client==2.2.4
pypdf
python-dotenv
flask